+++ src/chatgpt/exceptions/chatgptError.js
