+++ src/chatgpt/fhirToDocumentConverters/fhirToJsonDocumentConverter.js
@@ -1,5 +1,5 @@
 const {BaseFhirToDocumentConverter} = require('./baseFhirToDocumentConverter');
-const {ChatGPTDocument} = require('../chatgptDocument');
+const {ChatGPTDocument} = require('../structures/chatgptDocument');
 
 class FhirToJsonDocumentConverter extends BaseFhirToDocumentConverter {
     /**
