+++ src/chatgpt/structures/chatGPTResponse.js
