+++ src/chatgpt/structures/chatgptDocument.js
