+++ src/chatgpt/structures/chatgptMessage.js
