+++ src/tests/chatgpt/chatgpt.converters.test.js
@@ -8,8 +8,8 @@ dotenv.config({
 const patientBundleResource = require('./fixtures/patient.json');
 const patientCondensedBundleResource = require('./fixtures/patient_condensed.json');
 const {describe, test} = require('@jest/globals');
-const {ChatGPTManagerDirect} = require('../../chatgpt/chatgptManagerDirect');
-const {ChatGPTMessage} = require('../../chatgpt/chatgptMessage');
+const {ChatGPTManagerDirect} = require('../../chatgpt/managers/chatgptManagerDirect');
+const {ChatGPTMessage} = require('../../chatgpt/structures/chatgptMessage');
 const {FhirToJsonDocumentConverter} = require('../../chatgpt/fhirToDocumentConverters/fhirToJsonDocumentConverter');
 const {FhirToCsvDocumentConverter} = require('../../chatgpt/fhirToDocumentConverters/fhirToCsvDocumentConverter');
 const {FhirToSummaryDocumentConverter} = require('../../chatgpt/fhirToDocumentConverters/fhirToSummaryDocumentConverter');
