+++ src/tests/chatgpt/chatgpt.langchain.test.js
@@ -24,7 +24,7 @@ const {describe, test} = require('@jest/globals');
 const {MemoryVectorStore} = require('langchain/vectorstores/memory');
 const {Document} = require('langchain/document');
 const {ConsoleCallbackHandler} = require('langchain/callbacks');
-const {ChatGPTLangChainManager} = require('../../chatgpt/chatgptLangChainManager');
+const {ChatGPTLangChainManager} = require('../../chatgpt/managers/chatgptLangChainManager');
 const {FhirToSummaryDocumentConverter} = require('../../chatgpt/fhirToDocumentConverters/fhirToSummaryDocumentConverter');
 const {ResourceConverterFactory} = require('../../chatgpt/resourceConverters/resourceConverterFactory');
 const {createTestRequest, getTestContainer} = require('../common');
